migracion

+++ src/users/hooks/useLogin.tsx
@@ -8,14 +8,14 @@ interface HookProps extends LoginForm {
     register: boolean;
 }
 
-export default function useLogin({ name, email, password, register }: HookProps) {
+export default function useLogin({ username, email, password, register }: HookProps) {
     const navigate = useNavigate()
     const context = useContext(LoginContext)
 
     async function submit() {
         const emailValid = email.includes("@") && email.includes(".") && email.trim().length > 7
         const passwordValid = password.trim().length > 3
-        const nameValid = name.trim().length > 3
+        const nameValid = username.trim().length > 3
 
         if (register) {
             if (!emailValid || !passwordValid || !nameValid) {
@@ -30,14 +30,67 @@ export default function useLogin({ name, email, password, register }: HookProps)
                 return
             }
 
-            const res = await fetch("https://backend-gorras-app.vercel.app/users", {
+            const res = await fetch("https://gorras-backend-django-1ui3.vercel.app/users/create/", {
                 method: "POST",
                 headers: { "Content-Type": "application/json" },
-                body: JSON.stringify({ name, email, password, user_type: "client" })
+                body: JSON.stringify({ username, email, password})
             })
+            console.log(res.status)
+            if (res.status !== 201) {
+                Swal.fire({
+                    title: "Error",
+                    text: "El usuario ya existe.",
+                    icon: "error",
+                    background: "#292524",
+                    confirmButtonColor: "#C10007",
+                    color: "#fff"
+                })
+                return
+            }
             const data = await res.json()
-            localStorage.setItem("id", data._id)
+            const token = data.token
+
+
+            const profile = await fetch("https://gorras-backend-django-1ui3.vercel.app/profile/", {
+                method: "GET",
+                headers: { "Authorization": `Token ${token}` }
+            })
+            const profileData = await profile.json()
+            if(profile.status !== 200 ) {
+                Swal.fire({
+                    title: "Error",
+                    text: "Hubo un error al crear el usuario.",
+                    icon: "error",
+                    background: "#292524",
+                    confirmButtonColor: "#C10007",
+                    color: "#fff"
+                })
+                return;
+            }
+
+            const cartSession = await fetch("https://gorras-backend-django-1ui3.vercel.app/cart/cart_session/", {
+                method: "POST",
+                headers: { "Content-Type": "application/json" , "Authorization": `Token ${token}` },
+                body: JSON.stringify({
+                    user: await profileData.message.id,
+                })
+            })
+            const cartData = await cartSession.json()
+            localStorage.setItem("cartUser", cartData.id)
+            if(cartSession.status !== 201) {
+                Swal.fire({
+                    title: "Error",
+                    text: "Hubo un error al crear el usuario.",
+                    icon: "error",
+                    background: "#292524",
+                    confirmButtonColor: "#C10007",
+                    color: "#fff"
+                })
+                return;
+            }
             localStorage.setItem("auth", "true")
+            localStorage.setItem("Authorization", `Token ${token}`)
+            localStorage.setItem("id", JSON.stringify(profileData.message.id))
             context?.setLogged(true)
             Swal.fire({
                 title: "Usuario registrado",
@@ -49,7 +102,8 @@ export default function useLogin({ name, email, password, register }: HookProps)
             })
             navigate(-1)
 
-        } else {
+        }
+        else {
             if (!emailValid || !passwordValid) {
                 Swal.fire({
                     title: "Error",
@@ -62,11 +116,12 @@ export default function useLogin({ name, email, password, register }: HookProps)
                 return
             }
 
-            try {
-                const res = await fetch(`https://backend-gorras-app.vercel.app/users/login/${email}`)
-                const data = await res.json()
-
-                if (data.password !== password) {
+                const res = await fetch(`https://gorras-backend-django-1ui3.vercel.app/login/`, {
+                    method: "POST",
+                    headers: { "Content-Type": "application/json" },
+                    body: JSON.stringify({ email, password })
+                })
+                if(res.status !== 200){
                     Swal.fire({
                         title: "Error",
                         text: "Contraseña/email incorrectos, por favor intenta nuevamente.",
@@ -77,9 +132,16 @@ export default function useLogin({ name, email, password, register }: HookProps)
                     })
                     return
                 }
-
                 localStorage.setItem("auth", "true")
-                localStorage.setItem("id", data._id)
+                const data = await res.json()
+                localStorage.setItem("Authorization", `Token ${data.token}`)
+            const profile = await fetch("https://gorras-backend-django-1ui3.vercel.app/profile/", {
+                method: "GET",
+                headers: { "Authorization": `Token ${data.token}` }
+            })
+            const profileData = await profile.json()
+               localStorage.setItem("id", JSON.stringify(profileData.message.id))
+
                 context?.setLogged(true)
                 Swal.fire({
                     title: "Bienvenido de nuevo",
@@ -89,20 +151,19 @@ export default function useLogin({ name, email, password, register }: HookProps)
                     confirmButtonColor: "#00C950",
                     color: "#fff"
                 })
+                const cartSession = await fetch("https://gorras-backend-django-1ui3.vercel.app/cart/cart_session/", {
+                    method: "POST",
+                    headers: { "Content-Type": "application/json" , "Authorization": `Token ${data.token}` },
+                    body: JSON.stringify({
+                        user: await profileData.message.id,
+                    })
+                })
+                const cartData = await cartSession.json()
+                localStorage.setItem("cartUser", cartData.id)
                 navigate(-1)
 
-            } catch {
-                Swal.fire({
-                    title: "Error",
-                    text: "Contraseña/email incorrectos, por favor intenta nuevamente.",
-                    icon: "error",
-                    background: "#292524",
-                    confirmButtonColor: "#C10007",
-                    color: "#fff"
-                })
             }
-        }
     }
 
-    return { submit } // ✅ retorna la función
+    return { submit }
 }
